backend/requirements.txt
fastapi
uvicorn
pydantic
passlib[bcrypt]
PyJWT
joblib
numpy
scikit-learn
python-multipart
python-dotenv
langchain
huggingface_hub
langchain-huggingface
sqlmodel
